+++ src/VocabularyCoach.Services/Interfaces/IEditVocabularyService.cs
@@ -15,9 +15,11 @@ namespace VocabularyCoach.Services.Interfaces
 
 		Task<Uri> GetUrlForSpellCheck(LanguageText languageText, CancellationToken cancellationToken);
 
-		Task<LanguageText> AddLanguageText(LanguageTextCreationData languageTextData, CancellationToken cancellationToken);
+		Task<LanguageText> AddLanguageText(LanguageTextData languageTextData, CancellationToken cancellationToken);
 
 		Task<Translation> AddTranslation(LanguageText languageText1, LanguageText languageText2, CancellationToken cancellationToken);
+
+		Task<LanguageText> UpdateLanguageText(LanguageText languageText, LanguageTextData newLanguageTextData, CancellationToken cancellationToken);
 
 		Task DeleteLanguageText(LanguageText languageText, CancellationToken cancellationToken);