+++ src/VocabularyCoach.Services/Interfaces/Repositories/ILanguageTextRepository.cs
@@ -18,6 +18,8 @@ namespace VocabularyCoach.Services.Interfaces.Repositories
 
 		Task AddTranslation(Translation translation, CancellationToken cancellationToken);
 
+		Task UpdateLanguageText(LanguageText languageText, CancellationToken cancellationToken);
+
 		Task DeleteLanguageText(LanguageText languageText, CancellationToken cancellationToken);
 
 		Task DeleteTranslation(Translation translation, CancellationToken cancellationToken);