+++ src/VocabularyCoach/ViewModels/EditLanguageTextViewModel.cs
@@ -23,6 +23,8 @@ namespace VocabularyCoach.ViewModels
 {
 	public class EditLanguageTextViewModel : ObservableObject, IEditLanguageTextViewModel, INotifyDataErrorInfo
 	{
+		private readonly IVocabularyService vocabularyService;
+
 		private readonly IEditVocabularyService editVocabularyService;
 
 		private readonly IPronunciationRecordSynthesizer pronunciationRecordSynthesizer;
@@ -33,12 +35,31 @@ namespace VocabularyCoach.ViewModels
 
 		private readonly IMessenger messenger;
 
+		private bool newTextIsEdited;
+
+		public bool NewTextIsEdited
+		{
+			get => newTextIsEdited;
+			private set
+			{
+				SetProperty(ref newTextIsEdited, value);
+				OnPropertyChanged(nameof(ExistingTextIsEdited));
+			}
+		}
+
+		public bool ExistingTextIsEdited => !NewTextIsEdited;
+
 		public Language Language { get; private set; }
+
+		// This property is used only for 'edit text' mode.
+		private LanguageText EditedLanguageText { get; set; }
 
 		public bool RequireSpellCheck { get; private set; }
 
 		public bool CreatePronunciationRecord { get; private set; }
 
+		// This collection is filled both for 'new text' and for 'edit text' mode.
+		// For the later case, it is used for conflict detection.
 		public ObservableCollection<LanguageTextViewModel> ExistingTexts { get; } = new();
 
 		private bool textIsFocused;
@@ -90,6 +111,7 @@ namespace VocabularyCoach.ViewModels
 
 		private LanguageTextViewModel selectedText;
 
+		// This property is used only for 'new text' mode.
 		public LanguageTextViewModel SelectedText
 		{
 			get => selectedText;
@@ -104,6 +126,7 @@ namespace VocabularyCoach.ViewModels
 			}
 		}
 
+		// This property is always false for 'edit text' mode.
 		public bool ExistingTextIsSelected => SelectedText != null && SelectedText.TextWithNote == Text;
 
 		private string note;
@@ -120,6 +143,9 @@ namespace VocabularyCoach.ViewModels
 		}
 
 		private PronunciationRecord PronunciationRecord { get; set; }
+
+		// Is used in edit mode to detect change in pronunciation record;
+		private PronunciationRecord OriginalPronunciationRecord { get; set; }
 
 		private static IEnumerable<string> ValidationProperties
 		{
@@ -151,9 +177,11 @@ namespace VocabularyCoach.ViewModels
 
 		public ICommand ProcessEnterKeyCommand { get; }
 
-		public EditLanguageTextViewModel(IEditVocabularyService editVocabularyService, IPronunciationRecordSynthesizer pronunciationRecordSynthesizer,
-			IPronunciationRecordPlayer pronunciationRecordPlayer, IWebBrowser webBrowser, IMessenger messenger)
+		public EditLanguageTextViewModel(IVocabularyService vocabularyService, IEditVocabularyService editVocabularyService,
+			IPronunciationRecordSynthesizer pronunciationRecordSynthesizer, IPronunciationRecordPlayer pronunciationRecordPlayer,
+			IWebBrowser webBrowser, IMessenger messenger)
 		{
+			this.vocabularyService = vocabularyService ?? throw new ArgumentNullException(nameof(vocabularyService));
 			this.editVocabularyService = editVocabularyService ?? throw new ArgumentNullException(nameof(editVocabularyService));
 			this.pronunciationRecordSynthesizer = pronunciationRecordSynthesizer ?? throw new ArgumentNullException(nameof(pronunciationRecordSynthesizer));
 			this.pronunciationRecordPlayer = pronunciationRecordPlayer ?? throw new ArgumentNullException(nameof(pronunciationRecordPlayer));
@@ -165,17 +193,52 @@ namespace VocabularyCoach.ViewModels
 			ProcessEnterKeyCommand = new RelayCommand(() => messenger.Send(new EnterKeyPressedEventArgs()));
 		}
 
-		public async Task Load(Language language, bool requireSpellCheck, bool createPronunciationRecord, CancellationToken cancellationToken)
+		public async Task LoadForNewText(Language language, bool requireSpellCheck, bool createPronunciationRecord, CancellationToken cancellationToken)
+		{
+			NewTextIsEdited = true;
+
+			await Load(language, requireSpellCheck, createPronunciationRecord, cancellationToken);
+
+			EditedLanguageText = null;
+			OriginalPronunciationRecord = null;
+
+			ClearFilledData();
+		}
+
+		public async Task LoadForEditText(LanguageText editedLanguageText, bool requireSpellCheck, bool createPronunciationRecord, CancellationToken cancellationToken)
+		{
+			NewTextIsEdited = false;
+
+			await Load(editedLanguageText.Language, requireSpellCheck, createPronunciationRecord, cancellationToken);
+
+			EditedLanguageText = editedLanguageText;
+
+			Text = editedLanguageText.Text;
+			TextWasSpellChecked = true;
+
+			Note = editedLanguageText.Note;
+			SelectedText = null;
+
+			if (createPronunciationRecord)
+			{
+				PronunciationRecord = await vocabularyService.GetPronunciationRecord(editedLanguageText.Id, cancellationToken);
+			}
+
+			OriginalPronunciationRecord = PronunciationRecord;
+
+			ValidationIsEnabled = false;
+		}
+
+		private async Task Load(Language language, bool requireSpellCheck, bool createPronunciationRecord, CancellationToken cancellationToken)
 		{
 			Language = language;
 			RequireSpellCheck = requireSpellCheck;
 			CreatePronunciationRecord = createPronunciationRecord;
 
-			var existingTexts = await editVocabularyService.GetLanguageTexts(language, cancellationToken);
 			ExistingTexts.Clear();
-			ExistingTexts.AddRange(existingTexts.OrderBy(x => x.Text).Select(x => new LanguageTextViewModel(x)));
 
-			ClearFilledData();
+			var existingTexts = await editVocabularyService.GetLanguageTexts(language, cancellationToken);
+			ExistingTexts.AddRange(existingTexts.OrderBy(x => x.Text).Select(x => new LanguageTextViewModel(x)));
 		}
 
 		private async Task SpellCheckText(CancellationToken cancellationToken)
@@ -221,7 +284,17 @@ namespace VocabularyCoach.ViewModels
 
 		public async Task<LanguageText> SaveChanges(CancellationToken cancellationToken)
 		{
-			// For possible cases, see comment for method GetValidationErrorForText.
+			if (NewTextIsEdited)
+			{
+				return await SaveChangesForNewText(cancellationToken);
+			}
+
+			return await SaveChangesForEditText(cancellationToken);
+		}
+
+		private async Task<LanguageText> SaveChangesForNewText(CancellationToken cancellationToken)
+		{
+			// For possible cases, see comment for method GetValidationErrorForNewText.
 			if (ExistingTextIsSelected)
 			{
 				return SelectedText.LanguageText;
@@ -232,7 +305,7 @@ namespace VocabularyCoach.ViewModels
 				throw new InvalidOperationException("Pronunciation record is missing");
 			}
 
-			var textCreationData = new LanguageTextCreationData
+			var textData = new LanguageTextData
 			{
 				Language = Language,
 				Text = TextWithoutNote,
@@ -240,10 +313,30 @@ namespace VocabularyCoach.ViewModels
 				PronunciationRecord = PronunciationRecord,
 			};
 
-			var newText = await editVocabularyService.AddLanguageText(textCreationData, cancellationToken);
+			var newText = await editVocabularyService.AddLanguageText(textData, cancellationToken);
 			ExistingTexts.AddToSortedCollection(new LanguageTextViewModel(newText));
 
 			return newText;
+		}
+
+		private async Task<LanguageText> SaveChangesForEditText(CancellationToken cancellationToken)
+		{
+			if (CreatePronunciationRecord && PronunciationRecord == null)
+			{
+				throw new InvalidOperationException("Pronunciation record is missing");
+			}
+
+			var textData = new LanguageTextData
+			{
+				Language = Language,
+				Text = TextWithoutNote,
+				Note = Note,
+				PronunciationRecord = ReferenceEquals(PronunciationRecord, OriginalPronunciationRecord) ? null : PronunciationRecord,
+			};
+
+			var newLanguageText = await editVocabularyService.UpdateLanguageText(EditedLanguageText, textData, cancellationToken);
+
+			return newLanguageText;
 		}
 
 		public void ClearFilledData()
@@ -289,14 +382,14 @@ namespace VocabularyCoach.ViewModels
 
 			return propertyName switch
 			{
-				nameof(Text) => GetValidationErrorForText(),
+				nameof(Text) => NewTextIsEdited ? GetValidationErrorForNewText() : GetValidationErrorForEditText(),
 				_ => String.Empty,
 			};
 		}
 
-		private string GetValidationErrorForText()
+		private string GetValidationErrorForNewText()
 		{
-			/* Possible cases for text:
+			/* Possible cases:
 			 *   1. SelectedText == null && Text == String.Empty: User has neither typed nor picked any text - error.
 			 *   2. SelectedText == null && Text != String.Empty: User has typed new text.
 			 *   3. SelectedText != null && SelectedText.Text != Text: User has picked existing text and then edited it. We treat this as adding new text.
@@ -347,6 +440,37 @@ namespace VocabularyCoach.ViewModels
 
 			// We should not get here.
 			return "Internal error";
+		}
+
+		private string GetValidationErrorForEditText()
+		{
+			if (String.IsNullOrEmpty(TextWithoutNote))
+			{
+				return "Please type text";
+			}
+
+			if (RequireSpellCheck && !TextWasSpellChecked)
+			{
+				return "Please perform text spell check";
+			}
+
+			if (String.IsNullOrEmpty(Note))
+			{
+				if (ExistingTexts.Any(x => String.Equals(TextWithoutNote, x.Text, LanguageTextComparison.IgnoreCase)))
+				{
+					return $"Same text in {Language.Name} language already exists";
+				}
+			}
+			else
+			{
+				if (ExistingTexts.Any(x => String.Equals(TextWithoutNote, x.Text, LanguageTextComparison.IgnoreCase) &&
+				                           String.Equals(Note, x.Note, LanguageTextComparison.IgnoreCase)))
+				{
+					return $"Same text with the same note in {Language.Name} language already exists";
+				}
+			}
+
+			return String.Empty;
 		}
 	}
 }